Move functions specs into new "demo" project, as that's what they really are.
