+++ Solutions/Corvus.SpecFlow.Extensions.Demo/AzureFunctionsTesting/ScenariosUsingPerFeatureHookWithAdditionalConfiguration.feature.cs
@@ -10,7 +10,7 @@
 // ------------------------------------------------------------------------------
 #region Designer generated code
 #pragma warning disable
-namespace Corvus.SpecFlow.Extensions.Specs.Functions
+namespace Corvus.SpecFlow.Extensions.Demo.AzureFunctionsTesting
 {
     using TechTalk.SpecFlow;
     using System;
