+++ Solutions/Corvus.SpecFlow.Extensions.Demo/AzureFunctionsTesting/StepBindings.cs
@@ -1,4 +1,8 @@
-﻿namespace Corvus.SpecFlow.Extensions.Specs.Functions
+﻿// <copyright file="StepBindings.cs" company="Endjin Limited">
+// Copyright (c) Endjin Limited. All rights reserved.
+// </copyright>
+
+namespace Corvus.SpecFlow.Extensions.Demo.AzureFunctionsTesting
 {
     using System.Net;
     using System.Net.Http;
