+++ Solutions/Corvus.SpecFlow.Extensions.Demo/GlobalSuppressions.cs
@@ -1,0 +1,13 @@
+﻿// <copyright file="GlobalSuppressions.cs" company="Endjin Limited">
+// Copyright (c) Endjin Limited. All rights reserved.
+// </copyright>
+
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage(
+    "StyleCop.CSharp.OrderingRules",
+    "SA1210:Using directives should be ordered alphabetically by namespace",
+    Justification = "We need this until https://github.com/SpecFlowOSS/SpecFlow/issues/1828 is fixed")]
+
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage(
+    "StyleCop.CSharp.OrderingRules",
+    "SA1633:Using directives should be ordered alphabetically by namespace",
+    Justification = "We need this until https://github.com/SpecFlowOSS/SpecFlow/issues/1828 is fixed")]
