+++ Solutions/Corvus.SpecFlow.Extensions.Demo/specflow.json
@@ -1,0 +1,5 @@
+{
+  "stepAssemblies": [
+    { "assembly": "Corvus.SpecFlow.Extensions" }
+  ]
+}
