+++ Solutions/Corvus.SpecFlow.Extensions.sln
@@ -15,8 +15,15 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DevOps", "DevOps", "{B52766
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{1929703E-1336-4E35-B360-339B4AE41BDF}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.SpecFlow.Extensions.DemoFunction", "Corvus.SpecFlow.Extensions.DemoFunction\Corvus.SpecFlow.Extensions.DemoFunction.csproj", "{144715F8-4260-40AE-9FCE-78DAA5238AC3}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Demo", "Demo", "{C07A8D88-69A4-4ABC-A057-B09D55C6B61F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.SpecFlow.Extensions.Demo", "Corvus.SpecFlow.Extensions.Demo\Corvus.SpecFlow.Extensions.Demo.csproj", "{D70C5802-E069-4F7E-9342-DEEDD4F2502F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -36,9 +43,17 @@ Global
 		{144715F8-4260-40AE-9FCE-78DAA5238AC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{144715F8-4260-40AE-9FCE-78DAA5238AC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{144715F8-4260-40AE-9FCE-78DAA5238AC3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D70C5802-E069-4F7E-9342-DEEDD4F2502F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D70C5802-E069-4F7E-9342-DEEDD4F2502F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D70C5802-E069-4F7E-9342-DEEDD4F2502F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D70C5802-E069-4F7E-9342-DEEDD4F2502F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{144715F8-4260-40AE-9FCE-78DAA5238AC3} = {C07A8D88-69A4-4ABC-A057-B09D55C6B61F}
+		{D70C5802-E069-4F7E-9342-DEEDD4F2502F} = {C07A8D88-69A4-4ABC-A057-B09D55C6B61F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {00493C9B-EC8E-4A4A-B07B-26D6580B6CF6}
